feat: recreate index
