+++ pages/_document.tsx
@@ -12,11 +12,6 @@ export default class Document extends NextDocument {
           <link rel="stylesheet" type="text/css" charSet="UTF-8" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick.min.css" /> 
           <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick-theme.min.css" />
           <style id="stitches" dangerouslySetInnerHTML={{ __html: getCssText() }} />
-          <style>
-            .slick-next {
-                all: unset;
-            }
-          </style>
         </Head>
         <body>
           <Main />
